+++ cmd/client/domain.go
@@ -7,7 +7,7 @@ package main
 
 import (
 	"github.com/seeleteam/go-seele/contract/system"
-	"github.com/seeleteam/go-seele/rpc2"
+	"github.com/seeleteam/go-seele/rpc"
 )
 
 // createDomainName create a domain name
