+++ rpc/doc.go
