+++ rpc/socket.go
